moved mod files to their own folder
The mod files add a lot of clutter to the current simulation folder (especially when the .o and .c files are made). Moving them to a seperate folder does make loading the simulation slightly more annoything having to navigate to one folder than back out for the CNOW_run.hoc file, but its worth it considering how much cleaner the Current SImulation folder now is.