Removed unused files
The stackdump and .reg files were random data files that have no use (and haven't had a use for a long time).

The ResObj.hoc is no longer used since the removal of the ThreshFinderv1_2, but to give credit to the author it was moved to the Deprecated Code folder.