Add shebangs and mark python scripts as executable
Adding shebangs allows the python scripts to be run on Unix OSs like
regular executables instead of requiring the python interpreter to be
invoked directly. To allow this, the scripts also have to be marked as
executable. This marking is preserved by `git` via the mode setting and
is set when the repo is cloned on Unix machines. Windows ignores this
setting.