Separate out nrnmech.dll and CNOW_run.hoc
These files are loaded into Neuron when running simulations via Python.
The locations of these files was placed into `consts.py` and those
constants are now used in the scripts.

Note that there is also some special handling for the nrnmech.dll file.
On linux, this file is located in a different directory
('x86_64/.libs/') and is named 'libnrnmech.so' instead of 'nrnmech.dll'.
A simple check of `os.name` was added to check for this and then use the
proper path.