#56
Custom electrode configurations can now be exported using the saveElectrodeConfig() procedure. This creates a .ses file which can be loaded in NEURON and contains a series of HOC commands which will be run when the file is loaded.
It should be noted that this does not support saving of SineRamp or DCSine electrodes.