Fixed setting low_amp for custom square electrodes
The low_amp (or the anodic amplitude) was set to match the high_amp (or the cathodic amplitude). By multiplying the low_amp amplitude by -1 it now is the equal but opposite of the high_amp amplitude. This creates a charge balanced waveform as opposed to the DC current it was creating.