initial square wave gaps code
This is the basic experiment file that will likely later be multithreaded by passing different process id's. (see the README.md in the python scripts folder for more information)
The script sets up neuron, and then runs a series of block threshold calculations with different gap times. This can be edited to have gaps post anod and post cathod.